Solidified model
The model itself as well as the training and evaluation script were finalized and moved to separate .py files for easy access and use. A learning curve was plotted, seems like only ~5 epochs is a reasonable training length for this network with the used dataset size to avoid overfitting. 30 epoch and 5 epoch trained model weights are also included for easy tests on other machines.